policy-and-pagination

+++ app/Http/Controllers/PostController.php
@@ -11,8 +11,9 @@ class PostController extends Controller
 
     public function index()
     {
-        $posts = Post::all();
-        return view('admin.posts.index', ['posts' => $posts]);    
+        $posts = auth()->user()->posts()->paginate(5);
+
+        return view('admin.posts.index', ['posts' => $posts]);
     }
 
 
@@ -28,36 +29,44 @@ class PostController extends Controller
 
     public function store(Request $request)
     {
-
+        $this->authorize('create', Post::class);
 
         $inputs = $request->validate([
             'title' => ['required', 'min:8', 'max:255'],
-            'post_image' => ['file' ,'mimes:jpeg,png'],
+            'post_image' => ['file', 'mimes:jpeg,png'],
             'body' => ['required', 'min:8']
         ]);
-        
+
         if ($request->has('post_image')) {
             $inputs['post_image'] = 'storage/' . $request->post_image->store('images');
         }
 
         auth()->user()->posts()->create($inputs);
 
-        session()->flash('post-create-message', 'Post with title "'. $inputs['title'] .'" Created Successfuly');
+        session()->flash('post-create-message', 'Post with title "' . $inputs['title'] . '" Created Successfuly');
 
         return redirect()->route('post.index');
-
     }
 
     public function edit(Post $post)
     {
+        // $this->authorize('view', $post);
+
+        // if(auth()->user()->can('view', $post)) {
+        //     abort(403);
+        // }
+
         return view('admin.posts.edit', ['post' => $post]);
     }
 
     public function update(Post $post, Request $request)
     {
+        $this->authorize('update', $post);
+        
+
         $inputs = $request->validate([
             'title' => ['required', 'min:8', 'max:255'],
-            'post_image' => ['file' ,'mimes:jpeg,png'],
+            'post_image' => ['file', 'mimes:jpeg,png'],
             'body' => ['required', 'min:8']
         ]);
 
@@ -68,18 +77,19 @@ class PostController extends Controller
         $post->title = $inputs['title'];
         $post->body = $inputs['body'];
 
+
         // auth()->user()->posts()->save($post);
         $post->save();
 
-        session()->flash('post-update-message', 'Post with title "'. $inputs['title'] .'" Updated Successfuly');
+        session()->flash('post-update-message', 'Post with title "' . $inputs['title'] . '" Updated Successfuly');
 
         return redirect()->route('post.index');
-        
     }
 
 
     public function destroy(Post $post, Request $request)
     {
+        $this->authorize('delete', $post);
         $post->delete();
 
         // Session::flash('message', 'Post Was Deleted');
